feat(health): promote Coverage view into @repowise-dev/ui
Move the composed Coverage surface (summary metric cards, risk x coverage
scatter, untested-hotspot warnings, module rollup, per-file table with the
AI test-prompt action) into a shared CoverageView in packages/ui/src/health,
driven by the same CodeHealthAdapter. Web's coverage-tab.tsx is now a thin
host binding the adapter to its /api client and /repos/:id routing; output
is unchanged.
